Upgrade to PHPUnit 3.2.21

+++ testing/branches/2007-12-17/unittest/libraries/pear/PHPUnit/Extensions/PhptTestCase.php
@@ -39,16 +39,20 @@
  * @author     Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @copyright  2002-2008 Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @license    http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @version    SVN: $Id: PhptTestCase.php 1985 2007-12-26 18:11:55Z sb $
+ * @version    SVN: $Id: PhptTestCase.php 3165 2008-06-08 12:23:59Z sb $
  * @link       http://www.phpunit.de/
  * @since      File available since Release 3.1.4
  */
 
+if (PHPUnit_Util_Filesystem::fileExistsInIncludePath('PEAR/RunTest.php')) {
+    $currentErrorReporting = error_reporting(E_ERROR | E_WARNING | E_PARSE);
+    require_once 'PEAR/RunTest.php';
+    error_reporting($currentErrorReporting);
+}
+
 require_once 'PHPUnit/Framework.php';
 require_once 'PHPUnit/Extensions/PhptTestCase/Logger.php';
 require_once 'PHPUnit/Util/Filter.php';
-
-@include_once 'PEAR/RunTest.php';
 
 PHPUnit_Util_Filter::addFileToFilter(__FILE__, 'PHPUNIT');
 
@@ -60,20 +64,24 @@ PHPUnit_Util_Filter::addFileToFilter(__FILE__, 'PHPUNIT');
  * @author     Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @copyright  2002-2008 Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @license    http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @version    Release: 3.2.11
+ * @version    Release: 3.2.21
  * @link       http://www.phpunit.de/
  * @since      Class available since Release 3.1.4
  */
-class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
+class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test, PHPUnit_Framework_SelfDescribing
 {
     /**
      * The filename of the .phpt file.
      *
      * @var    string
-     * @access protected
      */
     protected $filename;
 
+    /**
+     * Options for PEAR_RunTest.
+     *
+     * @var    array
+     */
     protected $options = array();
 
     /**
@@ -82,7 +90,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * @param  string $filename
      * @param  array  $options Array with ini settings for the php instance run,
      *                         key being the name if the setting, value the ini value.
-     * @access public
      */
     public function __construct($filename, $options = array())
     {
@@ -111,7 +118,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * Counts the number of test cases executed by run(TestResult result).
      *
      * @return integer
-     * @access public
      */
     public function count()
     {
@@ -125,7 +131,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * @param  array $options Array with ini settings for the php instance run,
      *                        key being the name if the setting, value the ini value.
      * @return PHPUnit_Framework_TestResult
-     * @access public
      */
     public function run(PHPUnit_Framework_TestResult $result = NULL, $options = array())
     {
@@ -141,19 +146,19 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
             throw new InvalidArgumentException;
         }
 
-        $options = array_merge($options, $this->options);
-
         $coverage = $result->getCollectCodeCoverageInformation();
+        $options  = array_merge($options, $this->options);
 
         if ($coverage) {
-            $options = array('coverage' => TRUE);
+            $options['coverage'] = TRUE;
         } else {
-            $options = array();
+            $options['coverage'] = FALSE;
         }
 
-        $runner = new PEAR_RunTest(new PHPUnit_Extensions_PhptTestCase_Logger, $options);
+        $currentErrorReporting = error_reporting(E_ERROR | E_WARNING | E_PARSE);
+        $runner                = new PEAR_RunTest(new PHPUnit_Extensions_PhptTestCase_Logger, $options);
 
-        if ($coverage){
+        if ($coverage) {
             $runner->xdebug_loaded = TRUE;
         } else {
             $runner->xdebug_loaded = FALSE;
@@ -162,9 +167,9 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
         $result->startTest($this);
 
         PHPUnit_Util_Timer::start();
-        $buffer = $runner->run($this->filename, $options);
-        $time = PHPUnit_Util_Timer::stop();
-
+        $buffer       = $runner->run($this->filename, $options);
+        $time         = PHPUnit_Util_Timer::stop();
+        error_reporting($currentErrorReporting);
         $base         = basename($this->filename);
         $path         = dirname($this->filename);
         $coverageFile = $path . DIRECTORY_SEPARATOR . str_replace('.phpt', '.xdebug', $base);
@@ -175,11 +180,11 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
         $phpFile      = $path . DIRECTORY_SEPARATOR . str_replace('.phpt', '.php', $base);
 
         if (is_object($buffer) && $buffer instanceof PEAR_Error) {
-            $result->addError( 
-              $this, 
+            $result->addError(
+              $this,
               new RuntimeException($buffer->getMessage()),
-              $time 
-            ); 
+              $time
+            );
         }
 
         else if ($buffer == 'SKIPPED') {
@@ -205,16 +210,11 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
             }
         }
 
-        if ($coverage) {
+        if ($coverage && file_exists($coverageFile)) {
             eval('$coverageData = ' . file_get_contents($coverageFile) . ';');
             unset($coverageData[$phpFile]);
 
-            $codeCoverageInformation = array(
-              'test'  => $this,
-              'files' => $coverageData
-            );
-
-            $result->appendCodeCoverageInformation($this, $codeCoverageInformation);
+            $result->appendCodeCoverageInformation($this, $coverageData);
             unlink($coverageFile);
         }
 
@@ -227,7 +227,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * Returns the name of the test case.
      *
      * @return string
-     * @access public
      */
     public function getName()
     {
@@ -238,7 +237,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * Returns a string representation of the test case.
      *
      * @return string
-     * @access public
      */
     public function toString()
     {
